Start to use navigation(Nativetouter) and add new screen to display services and characteristics
