+++ example/package.json
@@ -17,7 +17,8 @@
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
     "react-native": "0.72.6",
-    "react-native-ble-manager": "file:../"
+    "react-native-ble-manager": "file:../",
+    "react-router-native": "^6.22.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
